// packages/engine/src/layout/layer-perspective-types.ts
/** Local X/Y rotation in degrees; camera distance in document pixels. */
export interface LayerRotation3d {
  x: number;
  y: number;
  perspective: number;
}
export interface LayerPerspectiveRect {
  x: number;
  y: number;
  width: number;
  height: number;
}
export interface LayerPerspectivePoint {
  x: number;
  y: number;
}